Rename taphome_sdk to legacy_taphome_sdk
