+++ legacy_taphome_sdk/operation_modes.py
