+++ legacy_taphome_sdk/taphome_api_service.py
