+++ legacy_taphome_sdk/thermostat_service.py
